Fixes provider name for XMLSignatureFactory/KeyInfoFactory to "XMLDSig"
This fix replaces the provider name "ApacheXMLDSig" used to instantiate
the XMLSignatureFactory/KeyInfoFactory by the string "XMLDSig". The
string "XMLDSig" makes sure, the old 1.4.x version of Apache XML
Security is directly loaded, instead of the new 2.0.x version in the
fallback. This is a problem, if both versions of XML Security are
available.